+++ src/main/java/model/ReverseRoute.java
@@ -7,7 +7,7 @@ public class ReverseRoute extends Route {
 
     private byte[] prevHop;
 
-    public ReverseRoute(byte[] destAddr, byte[] sourceAddr, byte hopCount, byte seq, byte[] prevHop, boolean validSeqNum) {
+    public ReverseRoute(byte[] destAddr, byte[] sourceAddr, byte hopCount, int seq, byte[] prevHop, boolean validSeqNum) {
         super(destAddr, sourceAddr, hopCount, seq, validSeqNum);
         this.prevHop = prevHop;
     }
