+++ src/main/java/packets/RREP.java
@@ -10,11 +10,11 @@ import java.util.Arrays;
 public class RREP extends Packet {
     private int lifetime;
     private byte[] destAddr;
-    private byte destSeqNum;
+    private int destSeqNum;
     private byte[] oriAddr;
     private byte hopCount;
 
-    public RREP(int lifetime, byte[] destAddr, byte destSeqNum, byte[] oriAddr, byte hopCount) {
+    public RREP(int lifetime, byte[] destAddr, int destSeqNum, byte[] oriAddr, byte hopCount) {
         super((byte) 2);
 
         this.lifetime = lifetime;
@@ -44,7 +44,7 @@ public class RREP extends Packet {
         return destAddr;
     }
 
-    public byte getDestSeqNum() {
+    public int getDestSeqNum() {
         return destSeqNum;
     }
 
@@ -61,7 +61,7 @@ public class RREP extends Packet {
     }
 
     public byte[] getBytes() {
-        var converted = Converter.prepareAddrPlusSeqNumToSend(new byte[]{destAddr[0], destAddr[1], destAddr[2], destAddr[3], destSeqNum,
+        var converted = Converter.prepareAddrPlusSeqNumToSend(new byte[]{destAddr[0], destAddr[1], destAddr[2], destAddr[3], (byte) destSeqNum,
                 oriAddr[0], oriAddr[1], oriAddr[2], oriAddr[3], hopCount});
 
         return Converter.prepareForEncoding(ArrayUtils.addAll(new byte[]{type, (byte) ((lifetime >> 12) & 0x3f), (byte) ((lifetime >> 6) & 0x3f), (byte) (lifetime & 0x3f)}, converted));
